+++ src/HRDNS/Socket/Server/Client.php
@@ -5,11 +5,20 @@ namespace HRDNS\Socket\Server;
 class Client
 {
 
+    /** @var string */
+    private $id = null;
+
     /** @var resource|null */
     private $socket = null;
 
     /** @var string */
-    private $id = null;
+    private $host = null;
+
+    /** @var integer */
+    private $port = null;
+
+    /** @var array */
+    private $attributes = array ();
 
     public function __construct()
     {
@@ -38,11 +47,70 @@ class Client
     }
 
     /**
+     * @param string $key
+     * @param mixed $value
+     * @return static
+     */
+    public function setAttribute($key, $value)
+    {
+        $this->attributes[$key] = $value;
+        return $this;
+    }
+
+    /**
+     * @param string $key
+     * @return mixed
+     */
+    public function getAttribute($key)
+    {
+        if (!isset($this->attributes[$key])) {
+            return null;
+        }
+        return $this->attributes[$key];
+    }
+
+    /**
      * @return resource|null
      */
     public function getSocket()
     {
         return $this->socket;
+    }
+
+    /**
+     * @param string $host
+     * @return static
+     */
+    public function setHost($host)
+    {
+        $this->host = $host;
+        return $this;
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getHost()
+    {
+        return $this->host;
+    }
+
+    /**
+     * @param integer $port
+     * @return static
+     */
+    public function setPort($port)
+    {
+        $this->port = (int)$port;
+        return $this;
+    }
+
+    /**
+     * @return integer|null
+     */
+    public function getPort()
+    {
+        return $this->port;
     }
 
 }